zavrsen user-story 3.13. Zavrseni svi user-story-ji. Nastavak:rad na sitnim korekcijama

+++ klinicki-centar/src/main/java/rs/ac/uns/ftn/informatika/jpa/dto/KlinikaDTO.java
@@ -8,7 +8,7 @@ public class KlinikaDTO {
 	private String naziv;
 	private String grad;
 	private String drzava;
-	private int ocena;
+	private double ocena;
 	private String adresa;
 	private String tip;
 	private KorisnikDTO admin;
@@ -65,12 +65,12 @@ public class KlinikaDTO {
 	}
 
 
-	public int getOcena() {
+	public double getOcena() {
 		return ocena;
 	}
 
 
-	public void setOcena(int ocena) {
+	public void setOcena(double ocena) {
 		this.ocena = ocena;
 	}
